refactor: extract config sound helpers
