+++ src/config.rs
@@ -1,12 +1,10 @@
-use std::path::PathBuf;
-
 use crossterm::event::{KeyCode, KeyModifiers};
 use serde::Deserialize;
 
 mod io;
 mod keybinds;
+mod sound;
 
-use self::io::resolve_config_relative_path;
 pub use self::{
     io::{
         config_dir, config_path, load_live_keybinds, save_onboarding_choices, upsert_section_bool,
@@ -16,6 +14,7 @@ pub use self::{
         format_key_combo, CommandKeybindConfig, CustomCommandAction, CustomCommandKeybind,
         Keybinds, LiveKeybindConfig,
     },
+    sound::{AgentSoundSetting, SoundConfig},
 };
 
 pub const CONFIG_PATH_ENV_VAR: &str = "HERDR_CONFIG_PATH";
@@ -23,9 +22,10 @@ pub const DEFAULT_SCROLLBACK_LIMIT_BYTES: usize = 10_000_000;
 use tracing::warn;
 
 #[cfg(test)]
-use self::{io::upsert_top_level_bool, keybinds::parse_key_combo};
+use std::path::PathBuf;
 
-use crate::detect::Agent;
+#[cfg(test)]
+use self::{io::upsert_top_level_bool, keybinds::parse_key_combo};
 
 pub fn app_dir_name() -> &'static str {
     io::app_dir_name()
@@ -170,127 +170,6 @@ pub struct AdvancedConfig {
     pub scrollback_limit_bytes: usize,
 }
 
-#[derive(Debug, Clone, Deserialize)]
-#[serde(default)]
-pub struct SoundConfig {
-    pub enabled: bool,
-    /// Optional mp3 file path used for all notification sounds.
-    /// Relative paths are resolved from the config file's directory.
-    pub path: Option<PathBuf>,
-    /// Optional mp3 file path for "done" notifications.
-    /// Relative paths are resolved from the config file's directory.
-    pub done_path: Option<PathBuf>,
-    /// Optional mp3 file path for "request" notifications.
-    /// Relative paths are resolved from the config file's directory.
-    pub request_path: Option<PathBuf>,
-    pub agents: AgentSoundOverrides,
-}
-
-#[derive(Debug, Clone, Deserialize)]
-#[serde(default)]
-pub struct AgentSoundOverrides {
-    pub pi: AgentSoundSetting,
-    pub claude: AgentSoundSetting,
-    pub codex: AgentSoundSetting,
-    pub gemini: AgentSoundSetting,
-    pub cursor: AgentSoundSetting,
-    pub cline: AgentSoundSetting,
-    pub open_code: AgentSoundSetting,
-    pub github_copilot: AgentSoundSetting,
-    pub kimi: AgentSoundSetting,
-    pub droid: AgentSoundSetting,
-    pub amp: AgentSoundSetting,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum AgentSoundSetting {
-    #[default]
-    Default,
-    On,
-    Off,
-}
-
-impl SoundConfig {
-    pub fn allows(&self, agent: Option<Agent>) -> bool {
-        if !self.enabled {
-            return false;
-        }
-
-        !matches!(self.agents.for_agent(agent), AgentSoundSetting::Off)
-    }
-
-    pub fn path_for(&self, sound: crate::sound::Sound) -> Option<PathBuf> {
-        let path = match sound {
-            crate::sound::Sound::Done => self.done_path.as_ref().or(self.path.as_ref()),
-            crate::sound::Sound::Request => self.request_path.as_ref().or(self.path.as_ref()),
-        }?;
-
-        Some(resolve_config_relative_path(path))
-    }
-
-    pub fn diagnostics(&self) -> Vec<String> {
-        let mut diagnostics = Vec::new();
-        for (field, path) in [
-            ("ui.sound.path", self.path.as_ref()),
-            ("ui.sound.done_path", self.done_path.as_ref()),
-            ("ui.sound.request_path", self.request_path.as_ref()),
-        ] {
-            let Some(path) = path else {
-                continue;
-            };
-
-            let resolved = resolve_config_relative_path(path);
-            if resolved
-                .extension()
-                .and_then(|ext| ext.to_str())
-                .is_none_or(|ext| !ext.eq_ignore_ascii_case("mp3"))
-            {
-                diagnostics.push(format!(
-                    "unsupported sound file format: {field} = {} resolves to {}; expected an mp3 file; using default sound",
-                    path.display(),
-                    resolved.display()
-                ));
-                continue;
-            }
-
-            if !resolved.exists() {
-                diagnostics.push(format!(
-                    "missing sound file: {field} = {} resolves to {}; using default sound",
-                    path.display(),
-                    resolved.display()
-                ));
-            } else if !resolved.is_file() {
-                diagnostics.push(format!(
-                    "invalid sound file: {field} = {} resolves to {}; using default sound",
-                    path.display(),
-                    resolved.display()
-                ));
-            }
-        }
-        diagnostics
-    }
-}
-
-impl AgentSoundOverrides {
-    pub fn for_agent(&self, agent: Option<Agent>) -> AgentSoundSetting {
-        match agent {
-            Some(Agent::Pi) => self.pi,
-            Some(Agent::Claude) => self.claude,
-            Some(Agent::Codex) => self.codex,
-            Some(Agent::Gemini) => self.gemini,
-            Some(Agent::Cursor) => self.cursor,
-            Some(Agent::Cline) => self.cline,
-            Some(Agent::OpenCode) => self.open_code,
-            Some(Agent::GithubCopilot) => self.github_copilot,
-            Some(Agent::Kimi) => self.kimi,
-            Some(Agent::Droid) => self.droid,
-            Some(Agent::Amp) => self.amp,
-            None => AgentSoundSetting::Default,
-        }
-    }
-}
-
 impl Default for KeysConfig {
     fn default() -> Self {
         Self {
@@ -339,41 +218,11 @@ impl Default for ToastConfig {
     }
 }
 
-impl Default for SoundConfig {
-    fn default() -> Self {
-        Self {
-            enabled: true,
-            path: None,
-            done_path: None,
-            request_path: None,
-            agents: AgentSoundOverrides::default(),
-        }
-    }
-}
-
 impl Default for AdvancedConfig {
     fn default() -> Self {
         Self {
             allow_nested: false,
             scrollback_limit_bytes: DEFAULT_SCROLLBACK_LIMIT_BYTES,
-        }
-    }
-}
-
-impl Default for AgentSoundOverrides {
-    fn default() -> Self {
-        Self {
-            pi: AgentSoundSetting::Default,
-            claude: AgentSoundSetting::Default,
-            codex: AgentSoundSetting::Default,
-            gemini: AgentSoundSetting::Default,
-            cursor: AgentSoundSetting::Default,
-            cline: AgentSoundSetting::Default,
-            open_code: AgentSoundSetting::Default,
-            github_copilot: AgentSoundSetting::Default,
-            kimi: AgentSoundSetting::Default,
-            droid: AgentSoundSetting::Off,
-            amp: AgentSoundSetting::Default,
         }
     }
 }
